[Refactor] All associations are the same!
I noticed `One` and `Many` are all the same, so I made them one.
